pyproject.toml
[tool.poetry]
name = "repl-nix-artisticgrowthassistant"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
openai = "^1.44.1"
pillow = "^10.4.0"
psycopg2-binary = "^2.9.9"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
